requirements.txt
accelerate==1.13.0
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cycler==0.12.1
Cython==3.2.4
diffusers==0.37.0
distro==1.9.0
easydict==1.13
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
ftfy==6.3.1
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
groq==1.1.2
openai==1.109.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnxruntime-gpu==1.23.2
open_clip_torch==3.3.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.7
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pillow==12.1.1
prettytable==3.17.0
protobuf==7.34.1
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.25
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
zipp==3.23.0
